requirements.txt
python-telegram-bot
python-dotenv
pillow
flask
flask-cors
requests
beautifulsoup4
lxml
gunicorn
